requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.13.0
numpy<2
geopandas==0.14.4
shapely==2.0.2
fiona==1.9.5
pillow==10.1.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
python-dotenv==1.0.0